+++ libpyclingo/clingo/tests/test_solving.py
@@ -196,6 +196,43 @@ class TestSolving(TestCase):
         self.ctl.ground([("base", [])])
         self.ctl.solve(on_model=on_model)
 
+    def test_cautious_consequences(self):
+        """
+        Test is_consequence function of model.
+        """
+
+        def lookup(m: Model, name: str):
+            return cast(SymbolicAtom, m.context.symbolic_atoms[Function(name)]).literal
+
+        def on_model(m: Model):
+            a = lookup(m, "a")
+            b = lookup(m, "b")
+            c = lookup(m, "c")
+            ca = m.is_consequence(a)
+            cb = m.is_consequence(b)
+            cc = m.is_consequence(c)
+            nca = m.is_consequence(-a)
+            ncb = m.is_consequence(-b)
+            ncc = m.is_consequence(-c)
+            self.assertTrue(ca is True)
+            self.assertTrue(nca is False)
+            self.assertTrue(ncb is False)
+            self.assertTrue(ncc is False)
+            if m.number == 1:
+                self.assertTrue(ncb is None or ncb is False)
+                self.assertTrue(ncc is None or ncc is False)
+                self.assertTrue(cb is None or cb is False)
+                self.assertTrue(cc is None or cc is False)
+                self.assertTrue(cb != cc)
+            if m.number == 2:
+                self.assertTrue(cb is False)
+                self.assertTrue(cc is False)
+
+        self.ctl.configuration.solve.enum_mode = "cautious"
+        self.ctl.add("base", [], "a. b | c.")
+        self.ctl.ground([("base", [])])
+        self.ctl.solve(on_model=on_model)
+
     def test_control_clause(self):
         """
         Test adding clauses while solving.